+++ src/rasterra/_array.py
@@ -537,6 +537,29 @@ class RasterArray(np.lib.mixins.NDArrayOperatorsMixin):
             new_data, self.transform, self._crs, no_data_value=self.no_data_value
         )
 
+    def select(
+        self,
+        x_coordinates: npt.NDArray[np.float64],
+        y_coordinates: npt.NDArray[np.float64],
+        method: str = "nearest",
+    ) -> npt.NDArray[DataDtypes]:
+        """Select values at specific coordinates."""
+        if x_coordinates.size != y_coordinates.size:
+            msg = "x and y coordinates must have the same size."
+            raise ValueError(msg)
+
+        if method == "nearest":
+            x_indices = np.searchsorted(self.x_coordinates(), x_coordinates)
+            y_indices = np.searchsorted(
+                self.y_coordinates(), y_coordinates, side="right"
+            )
+            # Flip y indices to match raster coordinates
+            y_indices = self.height - y_indices
+            return self._ndarray[y_indices, x_indices].copy()
+        else:
+            msg = "Only 'nearest' method is supported."
+            raise NotImplementedError(msg)
+
     def __repr__(self) -> str:
         out = "RasterArray\n"
         out += "===========\n"