+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Allow windowed reads of raster data.
+- Add a selection method for extracting values from a raster array by coordinates.
 
 ## [0.5.10] - 2024-05-15
 ### Added